requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.8
aiosignal==1.3.1
annotated-types==0.7.0
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
eval_type_backport==0.2.0
filelock==3.16.1
frozenlist==1.5.0
idna==3.10
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.1.0
numpy==2.1.3
pillow==10.4.0
propcache==0.2.0
pyarrow==18.1.0
pydantic==2.10.2
pydantic_core==2.27.1
Pygments==2.18.0
python-dotenv==1.0.1
requests==2.32.3
rich==13.9.4
shellingham==1.5.4
tabulate==0.9.0
together==1.3.5
tqdm==4.67.1
typer==0.13.1
typing_extensions==4.12.2
urllib3==2.2.3
yarl==1.18.0
